requirements.txt
python-telegram-bot>=21.0
openai>=1.0.0
python-docx>=1.0.0
mermaido>=0.2.0
python-dotenv>=1.0.0
playwright>=1.40.0

